pyproject.toml
[tool.poetry]
name = "hamatteru"
version = "0.1.0"
description = ""
authors = ["SORARAwo4649 <k11274649@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
django-debug-toolbar = "^3.2.1"
django-environ = "^0.4.5"
django-bootstrap-datepicker-plus = "^3.0.5"
django-bootstrap4 = "^3.0.1"
django-crispy-forms = "^1.11.2"
django-filter = "^2.4.0"
django-postgres-timestamp-without-tz = "^2021.1.15"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
